Neue Build-Scripte und Blatt3 jetzt mit GUI
