+++ Blatt3/pom.xml
@@ -19,11 +19,6 @@
 	</build>
 	<dependencies>
 		<dependency>
-			<groupId>org.jfree</groupId>
-			<artifactId>jfreechart</artifactId>
-			<version>1.0.19</version>
-		</dependency>
-		<dependency>
 			<groupId>de.buffalodan.ci</groupId>
 			<artifactId>CI-Library</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
