+++ CI-Library/src/de/buffalodan/ci/network/Connection.java
@@ -1,7 +1,5 @@
 package de.buffalodan.ci.network;
 
-import de.buffalodan.ci.network.neuron.Neuron;
-
 /**
  * @author Daniel
  *
