--- CI-Library/src/de/buffalodan/ci/network/neuron/InputNeuron.java
@@ -1,5 +1,0 @@
-package de.buffalodan.ci.network.neuron;
-
-public class InputNeuron {
-
-}
